backend/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
pandas==2.1.2
python-dotenv==1.2.1
slowapi==0.1.9
starlette==0.49.3
python-multipart==0.0.20
reportlab==4.4.4
fairlearn==0.13.0
scikit-learn==1.3.2
gspread==6.2.1
google-auth-oauthlib==1.3.1
aiohttp==3.13.2
jinja2==3.1.6
google-genai==1.47.0
rapidfuzz==3.13.0
